requirements.txt
piexif==1.1.3
Pillow==8.1.2
tqdm==4.59.0
win32-setctime==1.0.3
